[GUI] fix flyDialog's buggy recomputeSize method

+++ branches/avidemux_2.4_branch/avidemux/ADM_userInterfaces/ADM_QT4/ADM_dialog/T_flyDialog.cpp
@@ -51,7 +51,7 @@ protected:
 			UI_getPhysicalScreenSize(parent, &screenWidth, &screenHeight);
 			flyDialog->recomputeSize();
 			QCoreApplication::processEvents();
-			parent->move((screenWidth - parent->frameSize().width()) / 2, (screenHeight - parent->frameSize().height()) / 2);
+			parent->move((((int)screenWidth) - parent->frameSize().width()) / 2, (((int)screenHeight) - parent->frameSize().height()) / 2);
 		}
 		
 		return QObject::eventFilter(obj, event);
